requirements.txt
torch
transformers
pillow
pymupdf
numpy
faiss-cpu
langchain
langchain-community
scikit-learn
python-dotenv
streamlit
